integrated_projects/AI_CORE/data/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
paypalrestsdk==1.7.1
stripe==7.4.0
ccxt==4.0.0
openai==1.3.0
pillow==10.1.0
flask==3.0.0
flask-cors==4.0.0
sqlite3-python==1.0.0
sentry-sdk==1.38.0
cryptography==41.0.7
pydantic==2.5.0
aiohttp==3.9.1
